CodeCoverage: ReportGenerator moved to Generators\HtmlGenerator
