+++ Tester/tester.php
@@ -23,7 +23,7 @@ require __DIR__ . '/Framework/Assert.php';
 require __DIR__ . '/Framework/Dumper.php';
 require __DIR__ . '/Framework/DataProvider.php';
 require __DIR__ . '/Framework/TestCase.php';
-require __DIR__ . '/CodeCoverage/ReportGenerator.php';
+require __DIR__ . '/CodeCoverage/Generators/HtmlGenerator.php';
 
 
 $tester = new Tester\Runner\CliTester;
